requirements.txt
# Core dependencies
python-dotenv>=1.0.0
openai>=1.12.0
pandas>=2.1.0
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.66.0
tenacity>=8.2.0

# Analysis
scipy>=1.11.0
statsmodels>=0.14.0

# Optional: Jupyter for analysis
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0

# RL and NeMo dependencies - TESTED WORKING VERSIONS
# These specific versions are tested and compatible
# Install PyTorch first with: pip install torch==2.4.0 torchvision==0.19.0 --index-url https://download.pytorch.org/whl/cpu
transformers==4.44.0
peft==0.12.0
accelerate==0.33.0

# NeMo RL Framework (optional - script has fallback)
# Install with: pip install nemo-toolkit[nlp] nemo-aligner
# Note: May have import warnings but script uses fallback GRPO

# Additional RL utilities
datasets>=2.14.0
wandb>=0.15.0  # For experiment tracking (optional)
einops>=0.7.0

